compliance

+++ ShoppingItemRestApplication/src/main/java/de/maeddes/ShoppingItemRestApplication/ItemRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "shoppingItems", path = "v2")
-public interface ItemRepository extends JpaRepository<Item,Integer> {
+public interface ItemRepository extends JpaRepository<Item,Long> {
  
     List<Item> findByName(String name);
